+++ src/main/java/com/aurd/Student/Controller/courses/GetCourseDetailController.java
@@ -57,7 +57,7 @@ public class GetCourseDetailController {
 
 
         StudentCourseModel scm = studentCourseRepository.find("courseId=?1 and userId=?2",
-                courseEntity.getId(),request.getStud_id()).firstResult();
+                ((int) courseEntity.getId()),request.getStud_id()).firstResult();
 
         if(scm!=null){
             courseEntity.setPurchased(true);
